Cleaned up a bunch of code.
