+++ src/c#/main/command/miscellaneous/ToggleAutoWalkCommand.cs
