+++ src/c#/main/command/nation/NationCreateCommand.cs.meta
