+++ src/c#/main/command/nation/NationJoinCommand.cs.meta
